commands: move tapcli functionality to a new package
Relocate command-line parsing functionality to a new package
called `commands`. This change will eventually enable
integration tests to call into the parsing logic directly.
