+++ cmd/commands/conn.go
@@ -2,7 +2,7 @@
 // Copyright (c) 2015-2016 The Decred developers
 // Copyright (C) 2015-2022 The Lightning Network Developers
 
-package main
+package commands
 
 import (
 	"bufio"
