+++ cmd/commands/macaroon_jar.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"encoding/base64"
