+++ cmd/tapcli/main.go
@@ -4,12 +4,16 @@
 
 package main
 
-import "os"
+import (
+	"os"
+
+	"github.com/lightninglabs/taproot-assets/cmd/commands"
+)
 
 func main() {
 	// Set up the main CLI app.
-	app := NewApp()
+	app := commands.NewApp()
 	if err := app.Run(os.Args); err != nil {
-		Fatal(err)
+		commands.Fatal(err)
 	}
 }
